+++ Source/General/GlobalClass.vb
@@ -304,6 +304,12 @@ Public Class GlobalClass
             Log.WriteEnvironment()
             Log.WriteConfiguration()
 
+            Log.WriteHeader("Media Info Source File")
+            For Each i In p.SourceFiles
+                Log.WriteLine(i)
+            Next
+            Log.WriteLine(BR + MediaInfo.GetSummary(p.SourceFile))
+
             Log.WriteHeader($"{p.Script.Engine} Script")
             Log.WriteLine(p.Script.GetFullScript)