+++ Changelog.md
@@ -13,6 +13,7 @@ v2.xx.0 (not published yet)
 v2.30.0 (not published yet)
 ====================
 
+- Log: Fix not showing Media Info of source files in logs ([#1114](/../../issues/1114))
 - x265: Fix --log-level param values
 - Update tools
 - Update AviSynth+ plugins